act-adk-worker/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
sse-starlette>=2.1.0
google-genai>=1.0.0
google-cloud-firestore>=2.19.0
